% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_overview_module.R
\name{tibble_overview_module_ui}
\alias{tibble_overview_module_ui}
\alias{tibble_overview_server}
\title{tibble_overvie_module_ui and tibble_overview_module_server}
\usage{
tibble_overview_module_ui(id)

tibble_overview_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
present overview statistics about given tibble
}
\keyword{internal}
